chore: update podspec and update file name to Waypoint
