+++ Package.swift
@@ -1,3 +1,5 @@
+// swift-tools-version:5.3
+
 import PackageDescription
 
 let package = Package(
@@ -7,7 +9,7 @@ let package = Package(
         // Products define the executables and libraries a package produces, making them visible to other packages.
         .library(
             name: "waypoint",
-            targets: ["id"]),
+            targets: ["waypoint"]),
     ],
     targets: [
         // Targets are the basic building blocks of a package, defining a module or a test suite.
