--- Sources/id/Request.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-class Request {
-    var method: String
-    var params: [String: String]
-
-    init(method: String, params: [String: String]) {
-        self.method = method
-        self.params = params
-    }
-}
