--- Tests/id-packageTests/id_packageTests.swift
@@ -1,7 +1,0 @@
-import XCTest
-@testable import id_package
-
-final class id_packageTests: XCTestCase {
-    func testExample() throws {
-    }
-}
