+++ Tests/waypoint-packageTests/waypoint_packageTests.swift
@@ -1,0 +1,7 @@
+import XCTest
+@testable import waypoint
+
+final class waypoint_packageTests: XCTestCase {
+    func testExample() throws {
+    }
+}
